move FeatureBuilder into proper folder
