+++ SolastaCommunityExpansion/Classes/Tinkerer/TinkererClass.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using SolastaCommunityExpansion.Builders;
+using SolastaCommunityExpansion.Builders.Features;
 using SolastaCommunityExpansion.CustomFeatureDefinitions;
 using SolastaModApi;
 using UnityEngine;
